requirements.txt
argon2-cffi==20.1.0
asn1crypto==1.3.0
async-generator==1.10
attrs==20.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.2.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
gdown==3.12.2
glob2==0.7
idna==2.8
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
libarchive-c==2.9
lief==0.9.0
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.18.1
olefile==0.46
opencv-python==4.4.0.46
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.4
psutil==5.7.0
ptyprocess==0.6.0
pycocotools==2.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.22.0
ruamel-yaml==0.15.87
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
terminado==0.9.1
testpath==0.4.4
torch==1.5.0
torchvision==0.6.0a0+82fd1c8
tornado==6.1
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
